+++ HISTORY.md
@@ -5,10 +5,15 @@
 
 - Implemented functions `dot` (dot product), `cross` (cross product), and
   `nthRoot`.
-- Implemented recursive functions `clone`, `traverse`, `transform`, and `filter` 
-  for parsed node trees.
+- Opened up the API of nodes. Documented the API.
+- Implemented recursive functions `clone`, `traverse`, `transform`, and `filter`
+  for node trees.
 - Parameter `index` in the callbacks of `map` and `forEach` are now cloned
   for every callback.
+- Some internal refactoring inside nodes to make the API consistent:
+  - Renamed `params` to `args` and vice versa to make things consistent.
+  - Renamed `Block.nodes` to `Block.blocks`.
+  - `FunctionNode` now has a `name: string` instead of a `symbol: SymbolNode`.
 - Fixed `2e` giving a syntax error instead of being parsed as `2 * e`.